requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
opencv-python==4.5.3.56
dlib==19.22.1
torch==1.9.0
torchvision==0.10.0
numpy==1.21.2
pyserial==3.5
python-multipart==0.0.5
pydantic==1.8.2
gdown==4.4.0 